+++ src/content-script/api/private/stateTransitions/approveStateTransition.ts
@@ -45,7 +45,7 @@ export class ApproveStateTransitionHandler implements APIHandler {
     if (identity == null) {
       throw new Error(`Identity with identifier ${payload.identity} not found`)
     }
-    const stateTransition = await this.stateTransitionsRepository.get(payload.hash)
+    const stateTransition = await this.stateTransitionsRepository.getByHash(payload.hash)
 
     if (stateTransition == null) {
       throw new Error(`Could not find state transition with hash ${payload.hash} for signing`)
